Move modifiers and layer switch to oneshot keys
This is a substantial change and I'm going to have to actively
re-acclimate myself to it...

I went cold turkey and removed lgui and lctrl toggles.  These will be
the hardest for me.  I almost never use alt/meta... that change will be
comparatively easy.

(the branch name is a pun on a name of a videogame that I really ought
to finish: https://en.wikipedia.org/wiki/OneShot)